Echo'ing download links rather than mailing them out to all
Mailout to all@ss.com is too high impact given that
there's still a lot of fiddling involved in the build process.
The risk for spamming people with multiple release attempts is too high.
Also, we have a release mailinglist which people can subscribe to,
as well as a blog, etc, .etc.

The download links are typically required for sanity checks,
as well as pasting them into the release announcement.