Add composer.phar to the release downloads
In order to make it easier for newcomers to get started with
SilverStripe and subsequently start using Composer, this now bundles
composer.phar with the download. This means that that users can call
‘php composer.phar require bla’ without needing to go and figure out
how to install composer themselves.

A new version of composer.phar is downloaded every time phar archive is
run. To check that it’s compatible with SilverStripe, that version is
used to run the composer commands that these Phing actions rely on.